Update to the setup, change the UI for eaiser access
PlaySound and ChangeColor are the priamary parts of this update as they have gotten a new menu. Require components have been added to the scipts such that putting the script on a new file will put the required scripts on to the object automatically. They should also set their dependencies themself now.